Bug fix #33647
Particle combing didn't always respond as expected. The combed effect was 
depending on the center of object in view, instead of on the center of 
selected hairs. That made combing in certain close ups impossible.

Same was actually true for transform tools for hairs!

And even worse - there was an optimize break in calculating center
which wasn't true even. Causing centers for transform to fail similar
in other cases.




git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@53250 954f8c5b-7b00-dc11-b283-0030488c597c