Bugfix #33803
(2.65a and older)

Case: ALT+CTRL+U (user prefs) made mousewheel change button values in inactive
window.

On opening 2nd window, or de-activating a window, the modifier
state was kept for that window. That meant that mouse-over in a window was using
the old modifier - for example for scrollwheel it would change values
in buttons.




git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@53658 954f8c5b-7b00-dc11-b283-0030488c597c